Delete items linked to deleted characters

+++ src/server/router/routers/_characters.ts
@@ -1,5 +1,6 @@
 import { editCharacterSchema, newCharacterSchema } from "$src/types/zod-schema";
 import { z } from "zod";
+
 import { createProtectedRouter } from "../protected-router";
 
 export const protectedCharactersRouter = createProtectedRouter()
@@ -30,9 +31,30 @@ export const protectedCharactersRouter = createProtectedRouter()
 			id: z.string()
 		}),
 		async resolve({ input, ctx }) {
-			return await ctx.prisma.character.delete({
-				where: { id: input.id }
+			const character = await ctx.prisma.character.findUnique({
+				where: { id: input.id },
+				include: { logs: { include: { character: true } } }
 			});
+			if (character && character.userId !== ctx.session.user.id) {
+				const logIds = character.logs.map(log => log.id);
+				await ctx.prisma.magicItem.deleteMany({
+					where: {
+						logGainedId: {
+							in: logIds
+						}
+					}
+				});
+				await ctx.prisma.storyAward.deleteMany({
+					where: {
+						logGainedId: {
+							in: logIds
+						}
+					}
+				});
+				return await ctx.prisma.character.delete({
+					where: { id: input.id }
+				});
+			} else return false;
 		}
 	})
 	.query("getDMs", {
